:white_check_mark: [#288] Update/add tests
