+++ src/archiefbeheercomponent/destruction/tests/test_tasks.py
@@ -24,7 +24,6 @@ from archiefbeheercomponent.emails.tests.factories import AutomaticEmailFactory
 from archiefbeheercomponent.notifications.models import Notification
 from archiefbeheercomponent.report.constants import ReportTypeChoices
 from archiefbeheercomponent.report.models import DestructionReport
-from archiefbeheercomponent.report.tests.factories import DestructionReportFactory
 
 from ...constants import RoleTypeChoices
 from ...tests.utils import mock_service_oas_get
@@ -33,7 +32,6 @@ from ..models import ArchiveConfig, DestructionList, DestructionListItem
 from ..tasks import (
     check_if_reviewers_need_reminder,
     complete_and_notify,
-    create_destruction_zaak,
     process_destruction_list,
     process_list_item,
     update_zaak_from_list_item,
@@ -778,64 +776,3 @@ class UpdateZaakTests(TestCase):
         self.assertEqual(log.template, "destruction/logs/item_update_failed.html")
 
         mock_update_zaak.assert_called_once_with(list_item.zaak, archive_data, None)
-
-
-@requests_mock.Mocker()
-class CreateZaakTaskTests(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        Service.objects.create(
-            label="Zaken API",
-            api_type=APITypes.zrc,
-            api_root="https://oz.nl/zaken/api/v1/",
-            oas="https://oz.nl/zaken/api/v1/schema/openapi.json",
-        )
-        Service.objects.create(
-            label="Documenten API",
-            api_type=APITypes.drc,
-            api_root="https://oz.nl/documenten/api/v1/",
-            oas="https://oz.nl/documenten/api/v1/schema/openapi.json",
-        )
-
-    def _set_up_mocks(self, m):
-        mock_service_oas_get(
-            m,
-            "https://oz.nl/zaken/api/v1/",
-            "zrc",
-            oas_url="https://oz.nl/zaken/api/v1/schema/openapi.json",
-        )
-        mock_service_oas_get(
-            m,
-            "https://oz.nl/documenten/api/v1/",
-            "drc",
-            oas_url="https://oz.nl/documenten/api/v1/schema/openapi.json",
-        )
-        m.post(
-            "https://oz.nl/zaken/api/v1/zaken",
-            json={"url": "https://oz.nl/zaken/api/v1/zaken/123"},
-            status_code=201,
-        )
-        m.post(
-            "https://oz.nl/documenten/api/v1/enkelvoudiginformatieobjecten",
-            json={
-                "url": "https://oz.nl/documenten/api/v1/enkelvoudiginformatieobjecten/123"
-            },
-            status_code=201,
-        )
-        m.post("https://oz.nl/zaken/api/v1/zaakinformatieobjecten", status_code=201)
-        m.post("https://oz.nl/zaken/api/v1/resultaten", status_code=201)
-        m.post("https://oz.nl/zaken/api/v1/statussen", status_code=201)
-
-    def test_create_zaak_from_destruction_list(self, m):
-        self._set_up_mocks(m)
-
-        destruction_list = DestructionListFactory.create()
-        DestructionReportFactory.create(destruction_list=destruction_list)
-
-        create_destruction_zaak(destruction_list.id)
-
-        # Can't refresh from database due to fsm field
-        destruction_list = DestructionList.objects.get(id=destruction_list.id)
-        self.assertEqual(
-            "https://oz.nl/zaken/api/v1/zaken/123", destruction_list.zaak_url
-        )
